# docs/replicate-openapi.yaml
openapi: 3.0.0
info:
  title: Flow Editor - Replicate AI Integration API
  description: |
    Comprehensive API documentation for Replicate image editing and AI processing integration.
    This API provides endpoints to create, monitor, and manage AI-powered image predictions.
  version: 1.0.0
  contact:
    name: Flow Editor Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://app.example.com
    description: Production server
...
paths:
...
  /api/replicate/predictions/{id}:
    get:
      operationId: getPrediction
      tags:
        - Predictions
      summary: Get prediction status
      description: Retrieve the current status and results of a prediction
      parameters:
        - name: id
          in: path
          required: true
          description: Prediction ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Prediction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                status: succeeded
                output:
                  images:
                    - https://replicate.delivery/pbxt/abc123/result.jpg
                error: null
                created_at: '2024-10-20T10:30:00Z'
                completed_at: '2024-10-20T10:45:30Z'
        '400':
          description: Invalid prediction ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Prediction not found
                code: NOT_FOUND

    delete:
      operationId: cancelPrediction
      tags:
        - Predictions
      summary: Cancel prediction
      description: Cancel an in-progress prediction
      parameters:
        - name: id
          in: path
          required: true
          description: Prediction ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Prediction canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                status: canceled
        '404':
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Prediction:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique prediction identifier
        version:
          type: string
          description: Model version used
        status:
          type: string
          enum:
            - starting
            - processing
            - succeeded
            - failed
            - canceled
          description: Current prediction status
        input:
          type: object
          description: Input parameters sent to model
        output:
          type: object
          nullable: true
          description: Model output (null until completion)
          properties:
            images:
              type: array
              items:
                type: string
                format: uri
              description: URLs to generated images
        error:
          type: string
          nullable: true
          description: Error message if failed
        logs:
          type: string
          nullable: true
          description: Processing logs
        metrics:
          type: object
          properties:
            predict_time:
              type: number
              description: Processing time in seconds
        created_at:
          type: string
          format: date-time
          description: When prediction was created
        started_at:
          type: string
          format: date-time
          nullable: true
          description: When processing started
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: When processing finished
        urls:
          type: object
          properties:
            get:
              type: string
              format: uri
              description: URL to get prediction status
            cancel:
              type: string
              format: uri
              description: URL to cancel prediction
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code for programmatic handling
          enum:
            - INVALID_REQUEST
            - AUTHENTICATION_ERROR
            - RATE_LIMIT_ERROR
            - FILE_TOO_LARGE
            - PROCESSING_ERROR
            - TIMEOUT
            - NETWORK_ERROR
            - UNKNOWN_ERROR
            - CONFIGURATION_ERROR
            - VALIDATION_ERROR
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - SERVER_ERROR
            - BAD_GATEWAY
            - SERVICE_UNAVAILABLE
            - GATEWAY_TIMEOUT
        details:
          type: object
          nullable: true
          description: Additional error context
        requestId:
          type: string
          description: Request tracking ID
        timestamp:
          type: string
          format: date-time
          description: When error occurred
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - bearerAuth: []
...